# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TerraCare API",
    "version": "1.0.0",
    "description": "Terariumų, augintinių ir jų priežiūros valdymo API."
  },
  "servers": [
    {
      "url": "/",
      "description": "Dabartinis serveris"
    }
  ],
  "paths": {
    "/pets/{id}/profile": {
      "get": {
        "tags": [
          "Augintiniai"
        ],
        "summary": "Gauti augintinio profilį",
        "description": "Viename atsakyme sujungiama augintinio informacija, jo terariumas ir priežiūros istorija. Jei priežiūros įrašų nėra, care_records yra tuščias masyvas.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Augintinio ID.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Augintinio profilis iš trijų susijusių esybių",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "pet",
                    "terrarium",
                    "care_records"
                  ],
                  "properties": {
                    "pet": {
                      "$ref": "#/components/schemas/Pet"
                    },
                    "terrarium": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "temperature_min": {
                          "type": "number"
                        },
                        "temperature_max": {
                          "type": "number"
                        },
                        "humidity_min": {
                          "type": "number"
                        },
                        "humidity_max": {
                          "type": "number"
                        },
                        "last_cleaned": {
                          "type": "string",
                          "format": "date",
                          "nullable": true
                        },
                        "notes": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "care_records": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CareRecord"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Netinkamas augintinio ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Netinkamas augintinio ID."
                }
              }
            }
          },
          "404": {
            "description": "Augintinis nerastas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Augintinis nerastas."
                }
              }
            }
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nepavyko gauti augintinio profilio."
                }
              }
            }
          }
        }
      }
    },
    "/terrariums/{terrariumId}/pets/{petId}/care-records": {
      "get": {
        "tags": [
          "Priežiūros įrašai"
        ],
        "summary": "Gauti terariumo augintinio priežiūros įrašus",
        "description": "Tikrinama, ar terariumas egzistuoja ir ar augintinis priklauso būtent jam. Grąžinama augintinio priežiūros istorija nuo naujausių įrašų. Jei istorija tuščia, grąžinamas tuščias masyvas.",
        "parameters": [
          {
            "name": "terrariumId",
            "in": "path",
            "required": true,
            "description": "Terariumo ID.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          },
          {
            "name": "petId",
            "in": "path",
            "required": true,
            "description": "Šiam terariumui priklausančio augintinio ID.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Augintinio priežiūros įrašų sąrašas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CareRecord"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Netinkamas terariumo arba augintinio ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Terariumo ir augintinio ID turi būti teigiami sveikieji skaičiai."
                }
              }
            }
          },
          "404": {
            "description": "Terariumas nerastas arba augintinis šiame terariume nerastas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "terrariumNotFound": {
                    "summary": "Terariumas neegzistuoja",
                    "value": {
                      "error": "Terariumas nerastas."
                    }
                  },
                  "petNotInTerrarium": {
                    "summary": "Augintinis neegzistuoja arba priklauso kitam terariumui",
                    "value": {
                      "error": "Augintinis šiame terariume nerastas."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nepavyko gauti augintinio priežiūros įrašų."
                }
              }
            }
          }
        }
      }
    },
...
    "/care-records/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Priežiūros įrašo ID",
          "schema": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          }
        }
      ],
      "get": {
        "tags": [
          "Priežiūros įrašai"
        ],
        "summary": "Gauti priežiūros įrašą pagal ID",
        "responses": {
          "200": {
            "description": "Priežiūros įrašo informacija",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CareRecord"
                }
              }
            }
          },
          "400": {
            "description": "Netinkamas priežiūros įrašo ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Netinkamas priežiūros įrašo ID."
                }
              }
            }
          },
          "404": {
            "description": "Priežiūros įrašas nerastas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Priežiūros įrašas nerastas."
                }
              }
            }
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nepavyko gauti priežiūros įrašo."
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Priežiūros įrašai"
        ],
        "summary": "Atnaujinti priežiūros įrašą",
        "description": "Būtina pateikti pet_id, care_type ir notes. Įrašo laikas nekeičiamas.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CareRecordInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Priežiūros įrašas atnaujintas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CareRecord"
                }
              }
            }
          },
          "400": {
            "description": "Netinkamas ID arba netaisyklingas JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Netinkamas priežiūros įrašo ID."
                }
              }
            }
          },
          "404": {
            "$ref": "#/paths/~1care-records~1{id}/get/responses/404"
          },
          "422": {
            "$ref": "#/paths/~1care-records/post/responses/422"
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nepavyko atnaujinti priežiūros įrašo."
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Priežiūros įrašai"
        ],
        "summary": "Ištrinti priežiūros įrašą",
        "responses": {
          "204": {
            "description": "Priežiūros įrašas ištrintas. Atsakymo turinio nėra."
          },
          "400": {
            "$ref": "#/paths/~1care-records~1{id}/get/responses/400"
          },
          "404": {
            "$ref": "#/paths/~1care-records~1{id}/get/responses/404"
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nepavyko ištrinti priežiūros įrašo."
                }
              }
            }
          }
        }
      }
    },
    "/care-records": {
      "get": {
        "tags": [
          "Priežiūros įrašai"
        ],
        "summary": "Gauti priežiūros įrašų sąrašą",
        "description": "Grąžina vieną sąrašo puslapį nuo naujausių įrašų. Galima filtruoti pagal priežiūros tipą ir augintinį. Abu filtrai gali būti naudojami kartu. Nesant rezultatų, data yra tuščias masyvas.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Puslapio numeris nuo 1. Pernelyg didelis page ir limit derinys atmetamas su 400.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Įrašų skaičius puslapyje.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 5
            }
          },
          {
            "name": "care_type",
            "in": "query",
            "description": "Priežiūros įrašo tipas.",
            "schema": {
              "type": "string",
              "enum": [
                "feeding",
                "observation",
                "other"
              ]
            }
          },
          {
            "name": "pet_id",
            "in": "query",
            "description": "Augintinio ID. Jei atitinkančių įrašų nėra, grąžinamas tuščias sąrašas.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Priežiūros įrašų sąrašo puslapis su navigacijos nuorodomis",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "pagination",
                    "_links"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/CareRecord"
                          },
                          {
                            "type": "object",
                            "properties": {
                              "_links": {
                                "type": "object",
                                "properties": {
                                  "self": {
                                    "type": "string",
                                    "example": "/care-records/1"
                                  },
                                  "pet": {
                                    "type": "string",
                                    "example": "/pets/1"
                                  },
                                  "profile": {
                                    "type": "string",
                                    "example": "/pets/1/profile"
                                  }
                                }
                              }
                            }
                          }
                        ]
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "_links": {
                      "$ref": "#/components/schemas/PageLinks"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Netinkami puslapiavimo arba filtravimo parametrai",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Priežiūros tipas turi būti feeding, observation arba other."
                }
              }
            }
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nepavyko gauti priežiūros įrašų."
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Priežiūros įrašai"
        ],
        "summary": "Sukurti priežiūros įrašą",
        "description": "Įrašas priskiriamas egzistuojančiam augintiniui. Laikas nustatomas automatiškai.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CareRecordInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Priežiūros įrašas sukurtas",
            "headers": {
              "Location": {
                "description": "Sukurto priežiūros įrašo adresas",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CareRecord"
                }
              }
            }
          },
          "400": {
            "description": "Netaisyklingas JSON formatas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Netaisyklingas JSON formatas."
                }
              }
            }
          },
          "422": {
            "description": "Netinkami duomenys arba augintinis neegzistuoja",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nurodytas augintinis neegzistuoja."
                }
              }
            }
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nepavyko sukurti priežiūros įrašo."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 5
          },
          "total": {
            "type": "integer",
            "example": 12
          },
          "totalPages": {
            "type": "integer",
            "example": 3
          }
        }
      },
      "PageLinks": {
        "type": "object",
        "properties": {
          "self": {
            "type": "string"
          },
          "first": {
            "type": "string"
          },
          "previous": {
            "type": "string",
            "nullable": true
          },
          "next": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "CareRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "pet_id": {
            "type": "integer",
            "example": 1
          },
          "care_type": {
            "type": "string",
            "enum": [
              "feeding",
              "observation",
              "other"
            ]
          },
          "performed_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "description": "Laiką automatiškai nustato serveris."
          },
          "notes": {
            "type": "string",
            "example": "Augintinis aktyvus, pakitimų nepastebėta."
          }
        }
      },
      "CareRecordInput": {
        "type": "object",
        "required": [
          "pet_id",
          "care_type",
          "notes"
        ],
        "properties": {
          "pet_id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647,
            "example": 1
          },
          "care_type": {
            "type": "string",
            "enum": [
              "feeding",
              "observation",
              "other"
            ],
            "example": "observation"
          },
          "notes": {
            "type": "string",
            "minLength": 1,
            "description": "Pastabos negali būti sudarytos tik iš tarpų.",
            "example": "Augintinis aktyvus, pakitimų nepastebėta."
          }
        }
      },
      "Pet": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "terrarium_id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Pupa"
          },
          "species": {
            "type": "string",
            "example": "Litoria caerulea"
          },
          "arrival_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-09-01"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "example": "Aktyvi vakare."
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Nepavyko gauti augintinių sąrašo."
          }
        }
      }
    }
  }
}